Moving and improving types and references
