+++ extensions/docker-publisher/src/backend/index.ts
@@ -1,0 +1,5 @@
+import { ACRAPI } from './ACRAPI';
+import { DockerEngine } from './DockerEngine';
+import { DockerHubAPI } from './DockerHubAPI';
+
+export { ACRAPI, DockerEngine, DockerHubAPI };
