+++ extensions/docker-publisher/src/components/pages/RegistryType.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { useState, useEffect, useMemo } from 'react';
-import { OnChoiceDelegate } from '../../types/types';
+import { OnChoiceDelegate } from '../../types';
 import { ChooseRegistryAction } from '../ChooseRegistryAction';
 
 type Props = {
