+++ extensions/docker-publisher/src/components/pages/index.ts
@@ -2,7 +2,8 @@ import { PageRegistryType } from './RegistryType';
 import { DockerHubConfig } from './DockerHubConfig';
 import { RegistryConfig } from './RegistryConfig';
 import { ACRConfig } from './ACRConfig';
+import { ImageConfig } from './ImageConfig';
 import { Review } from './Review';
 import { Footer } from './Footer';
 
-export { PageRegistryType, DockerHubConfig, RegistryConfig, ACRConfig, Review, Footer };
+export { PageRegistryType, DockerHubConfig, RegistryConfig, ImageConfig, ACRConfig, Review, Footer };
